# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TourismPulseNZ API
  description: |
    TourismPulseNZ is a comprehensive tourism management platform for New Zealand, 
    providing smart recommendations, real-time analytics, demand forecasting, and 
    destination management capabilities.

    ## Features
    - Smart destination recommendations using ML algorithms
    - Real-time analytics and visitor tracking
    - Demand forecasting and peak season predictions
    - Weather-based destination filtering
    - Multi-role access control (Public, Tourist, Operator, Admin)
    - Booking and offer management

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

  version: 1.0.0
  contact:
    name: TourismPulseNZ Support
    email: support@tourismpulse.nz
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.tourismpulse.nz
    description: Production server
...
paths:
...
  /dest/api/v1/destinations:
    get:
      tags:
        - Destinations
      summary: Get all destinations
      description: Retrieve list of destinations with optional filtering
      operationId: getDestinations
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
            example: Adventure
        - name: region
          in: query
          description: Filter by region
          schema:
            type: string
            example: Queenstown
        - name: minRating
          in: query
          description: Minimum rating filter
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 5
            example: 4.0
        - name: maxPrice
          in: query
          description: Maximum price filter
          schema:
            type: number
            format: float
            example: 500
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: hiking,scenic
      responses:
        "200":
          description: Destinations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationsListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Destinations
      summary: Create destination (Admin only)
      description: Add a new destination to the system
      operationId: createDestination
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationCreate"
      responses:
        "201":
          description: Destination created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /dest/api/v1/destinations/{id}:
    get:
      tags:
        - Destinations
      summary: Get destination by ID
      description: Retrieve detailed information about a specific destination
      operationId: getDestinationById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Destination ID
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Destination retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationDetailResponse"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - Destinations
      summary: Update destination (Admin only)
      description: Update an existing destination
      operationId: updateDestination
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationUpdate"
      responses:
        "200":
          description: Destination updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Destinations
      summary: Delete destination (Admin only)
      description: Remove a destination from the system
      operationId: deleteDestination
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Destination deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Recommendation Endpoints
  /dest/api/recommendations:
    get:
      tags:
        - Recommendations
      summary: Get personalized recommendations
      description: Get smart destination recommendations based on user preferences and context
      operationId: getRecommendations
      security:
        - bearerAuth: []
      parameters:
        - name: topN
          in: query
          description: Number of recommendations to return
          schema:
            type: integer
            default: 10
            example: 5
        - name: categories
          in: query
          description: Preferred categories
          schema:
            type: string
            example: Adventure,Nature
        - name: tags
          in: query
          description: Preferred tags
          schema:
            type: string
        - name: minRating
          in: query
          description: Minimum rating
          schema:
            type: number
            format: float
        - name: maxPrice
          in: query
          description: Maximum price
          schema:
            type: number
            format: float
        - name: weatherCondition
          in: query
          description: Current weather condition
          schema:
            type: string
            enum: [sunny, rainy, cloudy, snowy]
        - name: season
          in: query
          description: Current season
          schema:
            type: string
            enum: [summer, autumn, winter, spring]
        - name: preferredRegions
          in: query
          description: Preferred regions
          schema:
            type: string
        - name: excludeVisited
          in: query
          description: Exclude previously visited destinations
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Recommendations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsResponse"
        "500":
          description: Failed to generate recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dest/api/recommendations/similar/{destinationId}:
    get:
      tags:
        - Recommendations
      summary: Get similar destinations
      description: Find destinations similar to a specific destination
      operationId: getSimilarDestinations
      parameters:
        - name: destinationId
          in: path
          required: true
          description: Destination ID to find similar destinations for
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Number of similar destinations to return
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Similar destinations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarDestinationsResponse"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dest/api/recommendations/trending:
    get:
      tags:
        - Recommendations
      summary: Get trending destinations
      description: Retrieve currently trending destinations based on recent activity
      operationId: getTrendingDestinations
      parameters:
        - name: limit
          in: query
          description: Number of trending destinations
          schema:
            type: integer
            default: 10
        - name: timeWindow
          in: query
          description: Time window for trending analysis
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 7d
      responses:
        "200":
          description: Trending destinations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendingDestinationsResponse"

  # Booking Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: float
          example: -45.0312
        longitude:
          type: number
          format: float
          example: 168.6626
    ContactInfo:
      type: object
      properties:
        phone:
          type: string
          example: "+64 3 249 7416"
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
    Destination:
      type: object
      properties:
        destination_id:
          type: integer
          example: 1
        name:
          type: string
          example: Milford Sound
        description:
          type: string
        category:
          type: string
          example: Nature & Wildlife
        region:
          type: string
          example: Southland
        location:
          $ref: "#/components/schemas/Location"
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.8
        estimated_cost:
          type: number
          format: float
          example: 250
        tags:
          type: array
          items:
            type: string
          example: [scenic, boat-tours, wildlife]
        images:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
          enum: [active, inactive, Open, Closed]
          example: active
        accessibility_features:
          type: boolean
        family_friendly:
          type: boolean
        pet_friendly:
          type: boolean
        opening_hours:
          type: string
        contact_info:
          $ref: "#/components/schemas/ContactInfo"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DestinationCreate:
      type: object
      required:
        - name
        - description
        - category
        - region
        - location
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        region:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        estimated_cost:
          type: number
          format: float
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        accessibility_features:
          type: boolean
        family_friendly:
          type: boolean
        pet_friendly:
          type: boolean
        opening_hours:
          type: string
        contact_info:
          $ref: "#/components/schemas/ContactInfo"
    DestinationUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        region:
          type: string
        estimated_cost:
          type: number
        status:
          type: string
          enum: [active, inactive]
        tags:
          type: array
          items:
            type: string
    DestinationsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          example: 15
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/Destination"
    DestinationDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        destination:
          $ref: "#/components/schemas/Destination"
    DestinationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        destination:
          type: object
          properties:
            destination_id:
              type: integer
            name:
              type: string
            status:
              type: string
            created_at:
              type: string
              format: date-time

    # Recommendation Schemas
    RecommendedDestination:
      allOf:
        - $ref: "#/components/schemas/Destination"
        - type: object
          properties:
            score:
              type: number
              format: float
              minimum: 0
              maximum: 1
              example: 0.92
            reasons:
              type: array
              items:
                type: string
              example:
                - Matches your preferred categories
                - Highly rated (4.7/5)
                - Perfect for current weather conditions
    RecommendationsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/RecommendedDestination"
    SimilarDestination:
      allOf:
        - $ref: "#/components/schemas/Destination"
        - type: object
          properties:
            similarity_score:
              type: number
              format: float
              example: 0.88
            reasons:
              type: array
              items:
                type: string
    SimilarDestinationsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        destinationId:
          type: integer
        count:
          type: integer
        similar:
          type: array
          items:
            $ref: "#/components/schemas/SimilarDestination"
    TrendingDestination:
      allOf:
        - $ref: "#/components/schemas/Destination"
        - type: object
          properties:
            trending_score:
              type: number
              format: float
              example: 0.95
            view_count:
              type: integer
              example: 12500
            booking_count:
              type: integer
              example: 450
            growth_rate:
              type: number
              format: float
              example: 2.3
            reasons:
              type: array
              items:
                type: string
    TrendingDestinationsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
        timeWindow:
          type: string
        trending:
          type: array
          items:
            $ref: "#/components/schemas/TrendingDestination"

    # Booking Schemas
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: object
...